refactor: remove unused MyForest (v1) code
