--- src/features/user/Profile/components/MyForestMap/microComponents/SingleMarkerImageContainer.tsx
@@ -1,45 +1,0 @@
-import {
-  PlantedTreesSvg,
-  RestoredSvg,
-  ConservationTreeSvg,
-} from '../../../../../../../public/assets/images/ProfilePageIcons';
-import theme from '../../../../../../theme/themeProperties';
-import MyForestMapStyle from '../../../styles/MyForestMap.module.scss';
-import { SingleMarkerUnitsProps } from './SingleMarkerUnits';
-
-type SingleMarkerImageContainerProps = Omit<SingleMarkerUnitsProps, 'units'>;
-
-const SingleMarkerImageContainer = ({
-  isNormalTreeDonation,
-  isRegisteredTree,
-  isRestorationTreePlantation,
-  isConservation,
-  isGiftContribution,
-  isMergeContribution,
-}: SingleMarkerImageContainerProps) => {
-  const { primaryColorNew, lightBlueColor } = theme;
-  const _renderTreeIcon = () => {
-    switch (true) {
-      case isNormalTreeDonation ||
-        isRegisteredTree ||
-        isMergeContribution ||
-        isGiftContribution:
-        return <PlantedTreesSvg color={`${primaryColorNew}`} />;
-      case isRestorationTreePlantation:
-        return <RestoredSvg color={`${primaryColorNew}`} />;
-      case isConservation:
-        return <ConservationTreeSvg color={`${lightBlueColor}`} />;
-      default:
-        return <></>;
-    }
-  };
-  //The condition for using the isMergeContribution scenario arises when a profile has two contributions,
-  //one categorized as a gift and the other as a regular tree donation with the same coordinates.
-  //In such a scenario, to prevent the overlap of a single marker on the map, the contributions are merged.
-  //This merging occurs when both contributions share the same projectId and are subsequently rendered using a cluster marker.
-  return (
-    <div className={MyForestMapStyle.svgContainer}>{_renderTreeIcon()}</div>
-  );
-};
-
-export default SingleMarkerImageContainer;
