--- src/features/user/Profile/styles/ProjectsContainer.module.scss
@@ -1,93 +1,0 @@
-@import '../../../../theme/theme';
-
-.tpoProjectsContainer {
-  // max-width: 1000px;
-  background-color: $backgroundBase;
-  margin-bottom: 80px;
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  @include smTabletView {
-    padding: 40px;
-    margin-bottom: 0px;
-  }
-}
-
-.projectsContainer {
-  width: 100%;
-  max-width: 1660px;
-  min-height: calc(100vh - 80px);
-  // display: flex;
-  // flex-direction: column;
-  // align-items: center;
-}
-
-.projectsTitleText {
-  font-family: $primaryFontFamily;
-  text-align: left;
-  width: 100%;
-  font-weight: 700;
-  margin-bottom: 20px;
-  color: $primaryFontColor;
-  font-size: $fontXXLarge;
-}
-
-//  the list of projects
-.listProjects {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: center;
-  padding: 0px 20px;
-}
-
-.singleProject {
-  width: 280px;
-  margin-top: 24px;
-  position: relative;
-}
-
-.projectNotFound {
-  font-family: $primaryFontFamily;
-  text-align: center;
-  padding: 0px 20px;
-  margin-top: 20px;
-  h2 {
-    margin-top: 8px;
-    font-weight: bold;
-    font-size: 24px;
-  }
-  &:hover {
-    cursor: pointer;
-  }
-}
-
-.projectNotFoundImage {
-  width: 100%;
-}
-
-@include smTabletView {
-  .listProjects {
-    padding: 0px;
-    justify-content: left;
-  }
-  .singleProject {
-    margin-right: 24px;
-  }
-}
-
-@include lgLaptopView {
-  .projectsTitleText {
-    text-align: left;
-  }
-  .listProjects {
-    justify-content: left;
-  }
-}
-
-@include smTabletView {
-  .listProjects {
-    text-align: left;
-  }
-}
