+++ src/features/user/TreeMapper/Analytics/components/Map/components/MapCredit/MapCredit.module.scss
@@ -1,0 +1,32 @@
+@import '../../../../../../../../theme/theme';
+
+.mapCreditsContainer {
+  position: absolute;
+  right: 92px;
+  bottom: 1px;
+  z-index: 1;
+  cursor: pointer;
+  @include xsPhoneView {
+    bottom: 51px;
+    right: 40px;
+  }
+  @include smTabletView {
+    bottom: 52px;
+    right: 48px;
+  }
+  @include mdTabletView {
+    bottom: 47px;
+    right: 93px;
+  }
+  @include lgLaptopView {
+    bottom: 47px;
+  }
+  @include xlDesktopView {
+    bottom: 0px;
+  }
+}
+
+.mapCredit {
+  padding: 10px;
+  width: 330px;
+}
