+++ src/features/user/TreeMapper/Analytics/components/Map/components/MapCredit/index.tsx
@@ -1,9 +1,9 @@
 import { useState, MouseEvent } from 'react';
 import { Typography, Popover } from '@mui/material';
-import myForestMapStyle from '../../../styles/MyForestMap.module.scss';
-import { InfoIcon } from '../../../../../../../public/assets/images/ProfilePageIcons/index';
+import styles from './MapCredit.module.scss';
+import { InfoIcon } from '../../../../../../../../../public/assets/images/ProfilePageIcons/index';
 
-const MyForestMapCredit = () => {
+const MapCredit = () => {
   const [anchorEl, setAnchorEl] = useState<HTMLButtonElement | null>(null);
   const handlePopoverOpen = (event: MouseEvent<HTMLButtonElement>) => {
     setAnchorEl(event.currentTarget);
@@ -15,7 +15,7 @@ const MyForestMapCredit = () => {
   const open = Boolean(anchorEl);
 
   return (
-    <div className={myForestMapStyle.mapCreditsContainer}>
+    <div className={styles.mapCreditsContainer}>
       <Typography
         aria-owns={open ? 'mouse-over-popover' : undefined}
         onMouseEnter={handlePopoverOpen}
@@ -37,7 +37,7 @@ const MyForestMapCredit = () => {
         onClose={handlePopoverClose}
         disableRestoreFocus
       >
-        <div className={myForestMapStyle.mapCredit}>
+        <div className={styles.mapCredit}>
           Esri Community Maps Contributors, Esri, HERE, Garmin, METI/NASA, USGS,
           Maxar, Earthstar Geographics, CNES/Airbus DS, USDA FSA, Aerogrid, IGN,
           IGP, and the GIS User Community
@@ -47,4 +47,4 @@ const MyForestMapCredit = () => {
   );
 };
 
-export default MyForestMapCredit;
+export default MapCredit;
