+++ src/server/router/_app.ts
@@ -1,8 +1,7 @@
 import { router } from '../trpc';
-import { myForestRouter, myForestV2Router } from './myForest';
+import { myForestV2Router } from './myForest';
 
 export const appRouter = router({
-  myForest: myForestRouter,
   myForestV2: myForestV2Router,
 });
